+++ src/contants.js
@@ -47,3 +47,8 @@ export const LIST_ITEM_CONTENTS = [
 export const SELECT_CATEGORY = ["한식", "중식", "일식", "양식", "아시안", "기타"];
 
 export const SELECT_DISTANCE = ["5분 내", "10분 내", "15분 내", "20분 내", "30분 내"];
+
+export const MODAL_BUTTONS_PROPERTY = [
+  { type: "button", stylingBased: "secondary", text: "취소하기" },
+  { type: "submit", stylingBased: "primary", text: "등록하기" },
+];
